requirements.txt
cloudpickle==1.6.0
json5==0.9.5
jsonschema==3.2.0
matplotlib==3.3.4
MLB-StatsAPI==1.2
numexpr==2.7.3
numpy==1.19.2
numpydoc==1.1.0
pandas==1.2.3
pandasql==0.7.3
pickleshare==0.7.5
py==1.10.0
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.9.0
tweepy==3.8.0